requirements.txt
absl-py==2.3.1
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
ffmpeg-python==0.2.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
future==1.0.0
humanfriendly==10.0
idna==3.11
itsdangerous==2.2.0
jax==0.8.0
jaxlib==0.8.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mediapipe==0.10.14
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnxruntime==1.23.2
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==12.0.0
polars==1.35.1
polars-runtime-32==1.35.1
protobuf==4.25.8
psutil==7.1.3
pycparser==2.23
pyodbc==5.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sounddevice==0.5.3
sympy==1.14.0
torch==2.9.0
torchvision==0.24.0
triton==3.5.0
typing_extensions==4.15.0
ultralytics==8.3.225
ultralytics-thop==2.0.18
urllib3==2.5.0
Werkzeug==3.1.3
